refactor: Refactor packages to have better and unified structure
