+++ packages/database/drizzle.config.ts
@@ -1,11 +1,12 @@
-import { databaseSchema, loadEnv } from '@terminal/env';
+import { loadEnv } from '@terminal/env';
+import { databaseSchema } from '@terminal/env/schemas';
 import { defineConfig } from 'drizzle-kit';
 
 const env = loadEnv(databaseSchema);
 
 export default defineConfig({
   out: './migrations',
-  schema: './schema',
+  schema: './src/schema',
   casing: 'snake_case',
   dialect: 'mysql',
   dbCredentials: {
