+++ packages/database/package.json
@@ -4,8 +4,9 @@
   "version": "0.0.0",
   "private": true,
   "exports": {
-    ".": "./index.ts",
-    "./migration": "./migration.ts"
+    ".": "./src/index.ts",
+    "./migration": "./src/migration.ts",
+    "./schema": "./src/schema/index.ts"
   },
   "scripts": {
     "db:generate": "drizzle-kit generate",
