+++ packages/database/src/index.ts
@@ -1,4 +1,4 @@
-import type { DatabaseSchema } from '@terminal/env';
+import type { DatabaseSchema } from '@terminal/env/schemas';
 import { drizzle } from 'drizzle-orm/mysql2';
 
 export function createDatabase(options: DatabaseSchema) {
@@ -8,5 +8,3 @@ export function createDatabase(options: DatabaseSchema) {
 }
 
 export type Database = ReturnType<typeof createDatabase>;
-
-export * from './schema';
