+++ packages/database/src/schema/auth.ts
